Feat: add prop for disabling the add button in the user profile page's address section
